requirements.txt
Django==4.2.11
gunicorn==21.2.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
djangorestframework==3.14.0
drf-yasg==1.21.7
setuptools>=68.0.0
django-ckeditor==6.7.0
Babel==2.13.1
djangorestframework_simplejwt==5.5.0
celery==5.5.0
redis==5.2.1
requests==2.32.3
Pillow==11.1.0
python-dotenv==1.0.1
drf-spectacular==0.27.2
django-filter==25.1
django-redis==5.4.0
geopy==2.4.1
django-ranged-response==0.2.0
django-rosetta==0.10.0
dj-database-url==2.3.0
google_play_scraper
packaging
jinja2
htmlmin
django-celery-beat
celery[redis]
fcm-django==2.0.0
firebase-admin==6.2.0
moviepy==1.0.3